Capture exception only if sentry is configured
